update files to minimize the code under build tags
